+++ src/contexts/ScheduleContext.tsx
@@ -7,8 +7,8 @@ import React, {
   useRef,
   useMemo,
 } from "react";
-import { Schedule } from "./types.ts";
-import dummyScheduleMap from "./dummyScheduleMap.ts";
+import { Schedule } from "../types.ts";
+import dummyScheduleMap from "../dummyScheduleMap.ts";
 
 interface ScheduleContextType {
   schedulesMap: Record<string, Schedule[]>;
